+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+ - Added DmcCoHd output card variant
+
 ## [13.1.1] - 2019-01-22
 ### Changed
  - Forcing NVX to use auto-initiation, disabling auto routing